// src/modules/garmentsProduction/pages/productionEntry/cutting.interface.ts
export type CuttingFormMaster = {
  floorId: string;
  lineId: string;
  buyerName: string;
  buyerId: string;
  styleId: string;
  styleNo: string;
  type: string;
  fabricType: string;
  layerNo: string;
  qty: number;
  cuttingQty: number;
  markerQty: number;
  layQty: number;
  bns: number;
  sns: number;
  remarks: string;
  minBundleQty: number;
  bundleQty: number;
  allowanceLength: number;
  markerWidth: number;
  markerLength: number;
  cuttingDate: string | null;
  cuttingNo: string;
};

export type CuttingFormDetails = {
  actions?: string;
  poNo: string;
  poId: number;
  colorId: number;
  colorName: string;
  lotBatchNo: string;
  rollNo: string;
  allowanceLength: number;
  allowanceWidth: number;
  utilized: number;
  wastage: number;
  uomId: number;
  uomName: string;
  shrinkage: string;
  gsm: string;
  length: number;
  weight: number;
  shade: string;
  layQty: number;
  rollBal: number;
  balance: number;
  actualBalance: number;
  placementMonth: string;
  deliveryDate: string;
};